placed src folder and data folder in model_training folder
