--- data/backup_raw/restaurant_reviews_without_rating.tsv
@@ -1,101 +1,0 @@
-Review
-Spend your money elsewhere.
-Their regular toasted bread was equally satisfying with the occasional pats of butter... Mmmm...!
-The Buffet at Bellagio was far from what I anticipated.
-And the drinks are WEAK, people!
--My order was not correct.
-Also, I feel like the chips are bought, not made in house.
-After the disappointing dinner we went elsewhere for dessert.
-The chips and sals a here is amazing!!!!!!!!!!!!!!!!!!!
-We won't be returning.
-This is my new fav Vegas buffet spot.
-I seriously cannot believe that the owner has so many unexperienced employees that all are running around like chickens with their heads cut off.
-Very, very sad.
-i felt insulted and disrespected, how could you talk and judge another human being like that?
-How can you call yourself a steakhouse if you can't properly cook a steak, I don't understand!
-I'm not impressed with the concept or the food.
-The only thing I wasn't too crazy about was their guacamole as I don't like it puréed.
-There is really nothing for me at postinos, hope your experience is better
-I got food poisoning here at the buffet.
-They brought a fresh batch of fries and I was thinking yay something warm but no!
-What SHOULD have been a hilarious, yummy Christmas Eve dinner to remember was the biggest fail of the entire trip for us.
-Needless to say, I won't be going back anytime soon.
-This place is disgusting!
-Every time I eat here, I see caring teamwork to a professional degree.
-The RI style calamari was a joke.
-However, there was so much garlic in the fondue, it was barely edible.
-I could barely stomach the meal, but didn't complain because it was a business lunch.
-It was so bad, I had lost the heart to finish it.
-It also took her forever to bring us the check when we asked for it.
-We aren't ones to make a scene at restaurants but I just don't get it...definitely lost the love after this one!
-Disappointing experience.
-The food is about on par with Denny's, which is to say, not good at all.
-If you want to wait for mediocre food and downright terrible service, then this is the place for you.
-WAAAAAAyyyyyyyyyy over rated is all I am saying.
-We won't be going back.
-The place was fairly clean but the food simply wasn't worth it.
-This place lacked style!!
-The sangria was about half of a glass wine full and was $12, ridiculous.
-Don't bother coming here.
-The meat was pretty dry, I had the sliced brisket and pulled pork.
-The building itself seems pretty neat, the bathroom is pretty trippy, but I wouldn't eat here again.
-It was equally awful.
-Probably not in a hurry to go back.
-very slow at seating even with reservation.
-Not good by any stretch of the imagination.
-The cashew cream sauce was bland and the vegetables were undercooked.
-The chipolte ranch dipping sause was tasteless, seemed thin and watered down with no heat.
-It was a bit too sweet, not really spicy enough, and lacked flavor.
-I was VERY disappointed!!
-This place is horrible and way overpriced.
-Maybe it's just their Vegetarian fare, but I've been twice and I thought it was average at best.
-It wasn't busy at all and now we know why.
-The tables outside are also dirty a lot of the time and the workers are not always friendly and helpful with the menu.
-The ambiance here did not feel like a buffet setting, but more of a douchey indoor garden for tea and biscuits.
-Con: spotty service.
-The fries were not hot, and neither was my burger.
-But then they came back cold.
-Then our food came out, disappointment ensued.
-The real disappointment was our waiter.
-My husband said she was very rude... did not even apologize for the bad food or anything.
-The only reason to eat here would be to fill up before a night of binge drinking just to get some carbs in your stomach.
-Insults, profound deuchebaggery, and had to go outside for a smoke break while serving just to solidify it.
-If someone orders two tacos don't' you think it may be part of customer service to ask if it is combo or ala cart?
-She was quite disappointed although some blame needs to be placed at her door.
-After all the rave reviews I couldn't wait to eat here......what a disappointment!
-Del Taco is pretty nasty and should be avoided if possible.
-It's NOT hard to make a decent hamburger.
-But I don't like it.
-Hell no will I go back
-We've have gotten a much better service from the pizza place next door than the services we received from this restaurant.
-"I don't know what the big deal is about this place, but I won't be back ""ya'all""."
-I immediately said I wanted to talk to the manager but I did not want to talk to the guy who was doing shots of fireball behind the bar.
-The ambiance isn't much better.
-Unfortunately, it only set us up for disapppointment with our entrees.
-The food wasn't good.
-Your servers suck, wait, correction, our server Heimer sucked.
-What happened next was pretty....off putting.
-too bad cause I know it's family owned, I really wanted to like this place.
-Overpriced for what you are getting.
-I vomited in the bathroom mid lunch.
-I kept looking at the time and it had soon become 35 minutes, yet still no food.
-I have been to very few places to eat that under no circumstances would I ever return to, and this tops the list.
-We started with the tuna sashimi which was brownish in color and obviously wasn't fresh.
-Food was below average.
-It sure does beat the nachos at the movies but I would expect a little bit more coming from a restaurant.
-All in all, Ha Long Bay was a bit of a flop.
-The problem I have is that they charge $11.99 for a sandwich that is no bigger than a Subway sub (which offers better and more amount of vegetables).
-Shrimp- When I unwrapped it (I live only 1/2 a mile from Brushfire) it was literally ice cold.
-It lacked flavor, seemed undercooked, and dry.
-It really is impressive that the place hasn't closed down.
-I would avoid this place if you are staying in the Mirage.
-The refried beans that came with my meal were dried out and crusty and the food was bland.
-Spend your money and time some place else.
-A lady at the table next to us found a live green caterpillar In her salad.
-the presentation of the food was awful.
-I can't tell you how disappointed I was.
-I think food should have flavor and texture and both were lacking.
-Appetite instantly gone.
-Overall I was not impressed and would not go back.
-The whole experience was underwhelming, and I think we'll just go to Ninja Sushi next time.
-Then, as if I hadn't wasted enough of my life there, they poured salt in the wound by drawing out the time it took to bring the check.
