--- data/raw/.gitignore
@@ -1,2 +1,0 @@
-/restaurant_reviews_without_rating.tsv
-/restaurant_reviews_with_rating.tsv
