+++ model_training/data/processed/.gitignore
@@ -1,0 +1,1 @@
+/c2_Classifier_Sentiment_Model
