+++ model_training/data/raw/.gitignore
@@ -1,0 +1,2 @@
+/restaurant_reviews_without_rating.tsv
+/restaurant_reviews_with_rating.tsv
