+++ model_training/src/predict.py
@@ -1,0 +1,5 @@
+"""Module to predict the sentiment of a sentence."""
+from models.predict_model import predict
+
+if __name__ == '__main__':
+    predict()
