+++ model_training/src/preprocess.py
@@ -1,0 +1,4 @@
+from data.preprocess import preprocess_data
+
+if __name__ == '__main__':
+    preprocess_data('model_training/data/raw/restaurant_reviews_with_rating.tsv')
