+++ setup.py
@@ -2,9 +2,9 @@ from setuptools import find_packages, setup
 
 setup(
     name='model_training',
-    version='0.1',
+    version='1.0.1',
     packages=find_packages(include=['model_training', 'model_training.*']),
-    package_data={'model_training': ['c1_BoW_Sentiment_Model.pkl', 'c2_Classifier_Sentiment_Model']},
+    package_data={'model_training/data/processed': ['model_training/data/processed/c1_BoW_Sentiment_Model.pkl', 'model_training/data/processed/c2_Classifier_Sentiment_Model']},
     include_package_data=True,
     install_requires=[
         'joblib==1.2.0',
