--- src/metrics.json
@@ -1,8 +1,0 @@
-{
-    "train": {
-        "accuracy": 0.7277777777777777,
-        "F1-Score": 0.7741935483870968,
-        "precision": 0.7058823529411765,
-        "recall": 0.8571428571428571
-    }
-}
