--- src/predict.py
@@ -1,5 +1,0 @@
-"""Module to predict the sentiment of a sentence."""
-from models.predict_model import predict
-
-if __name__ == '__main__':
-    predict()
