--- src/preprocess.py
@@ -1,4 +1,0 @@
-from data.preprocess import preprocess_data
-
-if __name__ == '__main__':
-    preprocess_data('data/raw/restaurant_reviews_with_rating.tsv')
